Make SearchPattern a top-level class.
